CNN/mac_requirements.txt
numpy
olefile
Pillow
PyYAML
six
torch 
torchvision 
